+++ management/model_enum_certificate_key_signagure_algorithm.go
@@ -15,7 +15,7 @@ import (
 	"fmt"
 )
 
-// EnumCertificateKeySignagureAlgorithm Specifies the signature algorithm of the key. Options are `SHA224withRSA`, `SHA256withRSA`, `SHA384withRSA` and `SHA512withRSA`.
+// EnumCertificateKeySignagureAlgorithm Specifies the signature algorithm of the key. For RSA keys, options are `SHA224withRSA`, `SHA256withRSA`, `SHA384withRSA` and `SHA512withRSA`. For elliptical curve (EC) keys, options are `SHA224with_ECDSA`, `SHA256with_ECDSA`, `SHA384with_ECDSA`, and `SHA512with_ECDSA`
 type EnumCertificateKeySignagureAlgorithm string
 
 // List of EnumCertificateKeySignagureAlgorithm
@@ -24,6 +24,10 @@ const (
 	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA256WITH_RSA EnumCertificateKeySignagureAlgorithm = "SHA256withRSA"
 	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA384WITH_RSA EnumCertificateKeySignagureAlgorithm = "SHA384withRSA"
 	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA512WITH_RSA EnumCertificateKeySignagureAlgorithm = "SHA512withRSA"
+	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA224WITH_ECDSA EnumCertificateKeySignagureAlgorithm = "SHA224with_ECDSA"
+	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA256WITH_ECDSA EnumCertificateKeySignagureAlgorithm = "SHA256with_ECDSA"
+	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA384WITH_ECDSA EnumCertificateKeySignagureAlgorithm = "SHA384with_ECDSA"
+	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA512WITH_ECDSA EnumCertificateKeySignagureAlgorithm = "SHA512with_ECDSA"
 )
 
 // All allowed values of EnumCertificateKeySignagureAlgorithm enum
@@ -32,6 +36,10 @@ var AllowedEnumCertificateKeySignagureAlgorithmEnumValues = []EnumCertificateKey
 	"SHA256withRSA",
 	"SHA384withRSA",
 	"SHA512withRSA",
+	"SHA224with_ECDSA",
+	"SHA256with_ECDSA",
+	"SHA384with_ECDSA",
+	"SHA512with_ECDSA",
 }
 
 func (v *EnumCertificateKeySignagureAlgorithm) UnmarshalJSON(src []byte) error {